API

Wrapper Functions

Most users will use the following wrapper functions, which automatically handles everything. The user only has to specify where the PLINK files (and possibly the phenotype/covariate files) are located.

MendelIHT.iht — Function

iht(plinkfile, k, kwargs...)

Runs IHT with sparsity level k. Example:

result = iht("plinkfile", 10)

Phenotypes and other covariates

Will use 6th column of .fam file for phenotype values and will automatically include an intercept as the only non-genetic covariate.

Arguments

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

k: An Int for sparsity parameter = number of none-zero coefficients

Optional Arguments

All arguments available in fit

iht(plinkfile, covariates, k, kwargs...)

Runs IHT with sparsity level k, with additional covariates stored separately. Example:

result = iht("plinkfile", "covariates.txt", 10)

Phenotypes

Will use 6th column of .fam file for phenotype values.

Other covariates

Covariates are read using readdlm function in Julia base. We require the covariate file to be comma separated, and not include a header line. Each row should be listed in the same order as in the PLINK. The first column should be all 1s to indicate an intercept. All other columns will be standardized to mean 0 variance 1.

Arguments

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

covariates: A String for covariate file name

k: An Int for sparsity parameter = number of none-zero coefficients

Optional Arguments

All arguments available in fit

iht(phenotypes, plinkfile, covariates, k, kwargs...)

Runs IHT with sparsity level k, where both phenotypes and additional covariates are stored separately. Example:

result = iht("phenotypes.txt", "plinkfile", "covariates.txt", 10)

Phenotypes

Phenotypes are read using readdlm function in Julia base. We require each subject's phenotype to occupy a different row. The file should not include a header line. Each row should be listed in the same order as in the PLINK.

Other covariates

Covariates are read using readdlm function in Julia base. We require the covariate file to be comma separated, and not include a header line. Each row should be listed in the same order as in the PLINK. The first column should be all 1s to indicate an intercept. All other columns will be standardized to mean 0 variance 1.

Arguments

phenotypes: A String for phenotype file name

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

covariates: A String for covariate file name

k: An Int for sparsity parameter = number of none-zero coefficients

Optional Arguments

All arguments available in fit

MendelIHT.cross_validate — Function

cross_validate(plinkfile, path, kwargs...)

Runs cross-validation to determinal optimal sparsity level k. Sparsity levels is specified in `path. Example:

mses = cross_validate("plinkfile", 1:20)
mses = cross_validate("plinkfile", [1, 2, 3, 4, 5]) # alternative syntax

Phenotypes and other covariates

Will use 6th column of .fam file for phenotype values and will automatically include an intercept as the only non-genetic covariate.

Arguments

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

path: Different sparsity levels. Can be an integer range (default 1:20) or vector of integers.

Optional Arguments

All arguments available in cv_iht

cross_validate(plinkfile, covariates, path, kwargs...)

Runs cross-validation to determinal optimal sparsity level k. Sparsity levels is specified in `path. Example:

mses = cross_validate("plinkfile", "covariates.txt", 1:20)
mses = cross_validate("plinkfile", "covariates.txt", [1, 10, 20]) # alternative syntax

Phenotypes

Will use 6th column of .fam file for phenotype values.

Other covariates

Covariates are read using readdlm function in Julia base. We require the covariate file to be comma separated, and not include a header line. Each row should be listed in the same order as in the PLINK. The first column should be all 1s to indicate an intercept. All other columns will be standardized to mean 0 variance 1.

Arguments

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

covariates: A String for covariate file name

path: Different sparsity levels. Can be an integer range (default 1:20) or vector of integers.

Optional Arguments

All arguments available in cv_iht

cross_validate(phenotypes, plinkfile, covariates, path, kwargs...)

Runs cross-validation to determinal optimal sparsity level k. Sparsity levels is specified in `path. Example:

mses = cross_validate("phenotypes.txt", "plinkfile", "covariates.txt", 1:20)
mses = cross_validate("phenotypes.txt", "plinkfile", "covariates.txt", [1, 10, 20]) # alternative syntax

Phenotypes

Phenotypes are read using readdlm function in Julia base. We require each subject's phenotype to occupy a different row. The file should not include a header line. Each row should be listed in the same order as in the PLINK.

Other covariates

Covariates are read using readdlm function in Julia base. We require the covariate file to be comma separated, and not include a header line. Each row should be listed in the same order as in the PLINK. The first column should be all 1s to indicate an intercept. All other columns will be standardized to mean 0 variance 1.

Arguments

phenotypes: A String for phenotype file name

plinkfile: A String for input PLINK file name (without .bim/.bed/.fam suffixes)

covariates: A String for covariate file name

path: Different sparsity levels. Can be an integer range (default 1:20) or vector of integers.

Optional Arguments

All arguments available in cv_iht

Core Functions

For advanced users, one can also run IHT regression or cross-validation directly. For cross validation, we generally recommend using cv_iht. This function cycles through the testing sets sequentially and fits different sparsity models in parallel. For larger problems (e.g. UK Biobank sized), one can instead choose to run cv_iht_distribute_fold. This function fits different sparsity models sequentially but initializes all training/testing model in parallel, which consumes more memory (see below). The later strategy allows one to distribute different sparsity parameters to different computers, achieving greater parallel power.

MendelIHT.cv_iht — Function

cv_iht(y, x, z; kwargs...)

For each model specified in path, performs q-fold cross validation and returns the (averaged) deviance residuals.

The purpose of this function is to find the best sparsity level k, obtained from selecting the model with the minimum out-of-sample error. Different cross validation folds are cycled through sequentially different paths are fitted in parallel on different CPUs. Currently there are no routines to cross validate different group sizes.

Warning

Do not remove files with random file names when you run this function. These are memory mapped files that will be deleted automatically once they are no longer needed.

Arguments

y: Response vector (phenotypes), should be an Array{T, 1}.

x: A design matrix (genotypes). Should be a SnpArray or an Array{T, 2}.

z: Matrix of non-genetic covariates of type Array{T, 2} or Array{T, 1}. The first column should be the intercept (i.e. column of 1).

Optional Arguments:

path: Different model sizes to be tested in cross validation (default 1:20)

q: Number of cross validation folds. (default 5)

d: Distribution of your phenotype. (default Normal)

l: A link function (default IdentityLink)

est_r: Symbol for whether to estimate nuisance parameters. Only supported distribution is negative binomial and choices include :Newton or :MM.

group: vector storing group membership for each predictor

weight: vector storing vector of weights containing prior knowledge on each predictor

folds: Vector that separates the sample into q disjoint subsets

destin: Directory where intermediate files will be generated. Directory name must end with /.

init: Boolean indicating whether we should initialize IHT algorithm at a good starting guess

use_maf: Boolean indicating we should scale the projection step by a weight vector

debias: Boolean indicating whether we should debias at each IHT step

verbose: Whether we want IHT to print meaningful intermediate steps

parallel: Whether we want to run cv_iht using multiple CPUs (highly recommended)

MendelIHT.cv_iht_distribute_fold — Function

Performs q-fold cross validation for Iterative hard thresholding to determine the best model size k. The function is the same as cv_iht except here each fold is distributed to a different CPU as opposed to each path to a different CPU.

This function has the edge over cv_iht because one can fit different sparsity levels on different computers. But this is assuming you have enough RAM and disk space to store all training data simultaneously.

StatsBase.fit — Function

fit(Mod::Type{<:StatisticalModel}, f::FormulaTerm, data, args...; 
    contrasts::Dict{Symbol}, kwargs...)

Convert tabular data into a numeric response vector and predictor matrix using the formula f, and then fit the specified model type, wrapping the result in a TableRegressionModel or TableStatisticalModel (as appropriate).

This is intended as a backstop for modeling packages that implement model types that are subtypes of StatsBase.StatisticalModel but do not explicitly support the full StatsModels terms-based interface. Currently this works by creating a ModelFrame from the formula and data, and then converting this to a ModelMatrix, but this is an internal implementation detail which may change in the near future.

Note

Do not delete intermediate files with random file names created by cv_iht and cv_iht_distribute_fold (windows users will be instructed to manually do so via print statements). These are memory-mapped files necessary for cross validation. For cv_iht, you must have x GB of free space and RAM on your hard disk where x is your .bed file size. For cv_iht_distribute_fold, you must have enough RAM and disk space to fit all q training datasets simultaneously, each of which typically requires (q - 1)/q * x GB.

Specifying Groups and Weights

When you have group and weight information, you input them as optional arguments in fit and cv_iht. The weight vector is a vector of Float64, while the group vector is a vector of Int. For instance,

    g = #import group vector
    w = #import weight vector
    ng = length(unique(g)) # specify number of non-zero groups
    result = fit(y, x, z; J=ng, k=10, d=Normal(), l=IdentityLink(), group=g, weight=w)

Simulation Utilities

MendelIHT provides some simulation utilities that help users explore the function and capabilities of iterative hard thresholding.

MendelIHT.simulate_random_snparray — Function

simulate_random_snparray(s::String, n::Integer, p::Integer; 
    [mafs::Vector{Float64}], [min_ma::Integer])

Creates a random SnpArray in the current directory without missing value, where each SNP has ⫺5 (default) minor alleles.

Note: if supplied minor allele frequency is extremely small, it could take a long time for the simulation to generate samples where at least min_ma (defaults to 5) are present.

Arguments:

s: name of SnpArray that will be created in the current directory. To not create file, use undef.

n: number of samples

p: number of SNPs

Optional Arguments:

mafs: vector of desired minor allele freuqencies (uniform(0,0.5) by default)

min_ma: the minimum number of minor alleles that must be present for each SNP (defaults to 5)

MendelIHT.simulate_correlated_snparray — Function

simulate_correlated_snparray(s, n, p; block_length, hap, prob)

Simulates a SnpArray with correlation. SNPs are divided into blocks where each adjacent SNP is the same with probability prob. There are no correlation between blocks.

Arguments:

n: number of samples

p: number of SNPs

s: name of SnpArray that will be created (memory mapped) in the current directory. To not memory map, use undef.

Optional arguments:

block_length: length of each LD block

hap: number of haplotypes to simulate for each block

prob: with probability prob an adjacent SNP would be the same.

Note

Simulating a SnpArray with $n$ subjects and $p$ SNPs requires up to $4np$ bits of RAM. Make sure you have enough RAM before simulating very large SnpArrays.

MendelIHT.simulate_random_response — Function

simulate_random_response(x, k, d, l; kwargs...)

This function simulates a random response (trait) vector y. When the distribution d is from Poisson, Gamma, or Negative Binomial, we simulate β ∼ N(0, 0.3) to roughly ensure the mean of response y doesn't become too large. For other distributions, we choose β ∼ N(0, 1).

Arguments

x: Design matrix

k: the true number of predictors.

d: The distribution of the simulated trait (note typeof(d) = UnionAll but typeof(d()) is an actual distribution: e.g. Normal)

l: The link function. Input canonicallink(d()) if you want to use the canonical link of d.

Optional arguments

r: The number of success until stopping in negative binomial regression, defaults to 10

α: Shape parameter of the gamma distribution, defaults to 1

Note

For negative binomial and gamma, the link function must be LogLink. For Bernoulli, the probit link seems to work better than logitlink when used in cv_iht or L0_reg.

MendelIHT.make_bim_fam_files — Function

make_bim_fam_files(x::SnpArray, y, name::String)

Creates .bim and .bed files from a SnpArray.

Arguments:

x: A SnpArray (i.e. .bed file on the disk) for which you wish to create corresponding .bim and .fam files.

name: string that should match the .bed file (Do not include .bim or .fam extensions in name).

y: Trait vector that will go in to the 6th column of .fam file.

Other Useful Functions

MendelIHT additionally provides useful utilities that may be of interest to a few advanced users.

MendelIHT.iht_run_many_models — Function

Runs IHT across many different model sizes specifed in path using the full design matrix. Same as cv_iht but DOES NOT validate in a holdout set, meaning that this will definitely induce overfitting as we increase model size. Use this if you want to quickly estimate a range of feasible model sizes before engaging in full cross validation.

MendelIHT.maf_weights — Function

maf_weights(x::SnpArray; max_weight::T = Inf)

Calculates the prior weight based on minor allele frequencies.

Returns an array of weights where w[i] = 1 / (2 * sqrt(p[i] (1 - p[i]))) ∈ (1, ∞). Here p is the minor allele frequency computed by maf() in SnpArrays.

x: A SnpArray

max_weight: Maximum weight for any predictor. Defaults to Inf.