Getting Started

Getting started

In this section, we outline the basic procedure to analyze your GWAS data with MendelIHT.

Installation

Press ] to enter package manager mode and type the following (after pkg>):

(v1.0) pkg> add https://github.com/OpenMendel/SnpArrays.jl
(v1.0) pkg> add https://github.com/biona001/MendelIHT.jl

The order of installation is important!

3 Step Workflow

Most analysis consists of 3 simple steps:

Import data.

Run cv_iht to determine best model size.

Run L0_reg to obtain final model.

We believe the best way to learn is through examples. Head over to the example section on the left to see these steps in action. Nevertheless, below contains function signatures and use cautions that any users should be aware.

Note

(1) MendelIHT assumes there are NO missing genotypes, and (2) the trios (.bim, .bed, .fam) must all be present in the same directory.

Core Functions

A standard analysis runs only 2 functions, other than importing data.

MendelIHT.cv_iht — Function.

cv_iht(d, l, x, z, y, J, path, num_fold)

Performs q-fold cross validation and returns the deviance residuals for each model.

Each CPU runs a different model for a given fold. To use this function, start julia using 4 (the more the better) processors by:

julia> using Distributed
julia> addprocs(4)

Arguments

d: A distribution (e.g. Normal, Bernoulli)

l: A link function (e.g. Loglink, ProbitLink)

x: A SnpArray, which can be memory mapped to a file. Does not engage in any linear algebra

z: Matrix of non-genetic covariates. The first column usually denotes the intercept.

y: Response vector

J: The number of maximum groups

path: Vector storing different model sizes

num_fold: Number of cross validation folds. For large data do not set this to be greater than 5

Optional Arguments:

group vector storing group membership

weight vector storing vector of weights containing prior knowledge on each SNP

folds: Vector that separates the sample into q disjoint subsets

init: Boolean indicating whether we should initialize IHT algorithm at a good starting guess

use_maf: Boolean indicating we should scale the projection step by a weight vector

debias: Boolean indicating whether we should debias at each IHT step

showinfo: Whether we want IHT to print meaningful intermediate steps

parallel: Whether we want to run cv_iht using multiple CPUs (highly recommended)

source

Note

Do not delete intermediate files (e.g. train.bed) created by cv_iht. These are memory-mapped files necessary for cross validation. In order to successfully create these files, you must have x GB of free space on your hard disk where x is your .bed file size.

MendelIHT.L0_reg — Function.

L0_reg(x, xbm, z, y, J, k, d, l)

Runs Iterative Hard Thresholding for GWAS data x, response y, and non-genetic covariates z on a specific sparsity parameter k.

One needs to construct a SnpBitMatrix type (xbm below) before running this function.

Arguments:

x: A SnpArray, which can be memory mapped to a file. Does not engage in any linear algebra

xbm: The bitarray representation of x. This matrix is loaded in RAM and performs linear algebra. It's possible to set scale=false for xbm, especially when rare SNPs exist

z: Matrix of non-genetic covariates. The first column usually denotes the intercept.

y: Response vector

J: The number of maximum groups

k: Number of non-zero predictors in each group

d: A distribution (e.g. Normal, Poisson)

l: A link function (e.g. Loglink, ProbitLink)

Optional Arguments:

group vector storing group membership

weight vector storing vector of weights containing prior knowledge on each SNP

use_maf indicates whether we want to scale the projection with minor allele frequencies (see paper)

debias is boolean indicating whether we debias at each iteration (see paper)

show_info boolean indicating whether we want to print results if model does not converge. Should set to false for multithread/multicore computing

init boolean indicating whether we want to initialize β to sensible values through fitting. This is not efficient yet.

tol is used to track convergence

max_iter is the maximum IHT iteration for a model to converge. Defaults to 200, or 100 for cross validation

max_step is the maximum number of backtracking. Since l0 norm is not convex, we have no ascent guarantee

source

Supported GLM models and Link functions

MendelIHT borrows distribution and link functions implementationed in GLM.jl and Distributions.jl

Distributions (listed with their canonical link) that work with L0_reg and cv_iht are:

          Normal (IdentityLink)
       Bernoulli (LogitLink)
         Poisson (LogLink)
NegativeBinomial (LogLink)
           Gamma (InverseLink) **(not tested)**
 InverseGaussian (InverseSquareLink) **(not tested)**

Available link functions are:

CauchitLink
CloglogLink
IdentityLink
InverseLink
InverseSquareLink
LogitLink
LogLink
ProbitLink
SqrtLink

Tip

For logistic regression, the ProbitLink seems to work better than LogitLink. For d = NegativeBinomial or d=Gamma, the link function must be LogLink.

Specifying Groups and Weights

When you have group and weight information, you input them as optional arguments in L0_reg and cv_iht. The weight vector is a vector of Float64, while the group vector is a vector of integers. For instance,

    g = #import group vector
    w = #import weight vector
    result = L0_reg(x, xbm, z, y, 1, k, d(), l, group=g, weight=w)

Simulation Utilities

MendelIHT provides some simulation utilities that help users explore the function and capabilities of iterative hard thresholding.

MendelIHT.simulate_random_snparray — Function.

simulate_random_snparray(n::Integer, p::Integer, s::String; mafs::Vector{Float64}, min_ma::Integer)

Creates a random SnpArray in the current directory without missing value, where each SNP has ⫺5 (default) minor alleles.

Note: if supplied minor allele frequency is extremely small, it could take a long time for the simulation to generate samples where at least min_ma (defaults to 5) are present.

Arguments:

n: number of samples

p: number of SNPs

s: name of the simulated SnpArray that will be created on the current directory

Optional Arguments:

mafs: vector of desired minor allele freuqencies

min_ma: the minimum number of minor alleles that must be present for each SNP (defaults to 5)

source

Note

Simulating a SnpArray with $n$ subjects and $p$ SNPs requires roughly $n \times p \times 4$ bits of RAM. Make sure you have enough RAM before simulating very large SnpArrays.

MendelIHT.simulate_random_response — Function.

simulate_random_response(x::SnpArray, xbm::SnpBitMatrix, k::Int, d::UnionAll, l::Link)

This function simulates a random response (trait) vector y based on provided x, β, distirbution, and link function.

When the distribution is from Poisson, Gamma, or Negative Binomial, we simulate β ∼ N(0, 0.3) to roughly ensure the mean of response y doesn't become too large. For other distributions, we choose β ∼ N(0, 1).

Arguments

x: The SnpArray

xbm: SnpBitMatrix type of your SnpArray

k: the true number of predictors.

d: The distribution of the simulated trait (note typeof(d) = UnionAll but typeof(d()) is an actual distribution: e.g. Normal)

l: The link function. Input canonicallink(d()) if you want to use the canonical link of d.

Optional arguments

nn: The number of success until stopping in negative binomial regression, defaults to 10

α: Shape parameter of the gamma distribution, defaults to 1

source

Note

For negative binomial and gamma, the link function must be LogLink. For Bernoulli, the probit link seems to work better than logitlink when used in cv_iht or L0_reg.

MendelIHT.adhoc_add_correlation — Function.

adhoc_add_correlation(x::SnpArray, ρ::Float64, pos::Int64, location::Vector{Int})

Makes 1 SNP (a column of x) correlate with SNPs in location with correlation coefficient roughly ρ.

Arguments:

x: the snparray

ρ: correlation coefficient

pos: the position of the target SNP that everything would be correlated to

location: All the SNPs that shall be correlated with the SNP at position pos with correlation ρ.

source

MendelIHT.make_bim_fam_files — Function.

make_bim_fam_files(x::SnpArray, y, name::String)

Creates .bim and .bed files from a SnpArray.

Arguments:

x: A SnpArray (i.e. .bed file on the disk) for which you wish to create corresponding .bim and .fam files. name: string that should match the .bed file (Do not include .bim or .fam extensions in name). y: Trait vector that will go in to the 6th column of .fam file.

source

Other Useful Functions

MendelIHT additionally provides useful utilities that may be of interest to a few advanced users.

MendelIHT.iht_run_many_models — Function.

Runs IHT across many different model sizes specifed in path.

This is basically the same as cv_iht except we DO NOT validate each model in a holdout set, meaning that this will definitely induce overfitting as we increase model size. Use this to perform a quick estimate a range of feasible model sizes before engaging in full cross validation.

source

MendelIHT.loglikelihood — Function.

loglikelihood(y::AbstractVector, xb::AbstractVector, d::UnivariateDistribution)

This function calculates the loglikelihood of observing y given μ = g^{-1}(xb) and some distribution d from the exp family. Note that loglikelihood is the sum of the logpdfs for each observation. For each logpdf, we

source

MendelIHT.project_k! — Function.

project_k!(x::AbstractVector, k::Integer)

This function projects a vector x onto the set Sk = { y in R^p : || y ||0 <= k }.

Arguments:

b is the vector to project.

k is the number of components of b to preserve.

source

MendelIHT.project_group_sparse! — Function.

project_group_sparse!(y::AbstractVector, group::AbstractVector, J::Integer, k::Integer)

Projects the vector y onto the set with at most J active groups and at most k active predictors per group. The variable group encodes group membership. Currently assumes there are no unknown or overlaping group membership.

source

MendelIHT.maf_weights — Function.

maf_weights(x::SnpArray; max_weight::T = Inf)

Calculates the prior weight based on minor allele frequencies.

Returns an array of weights where w[i] = 1 / (2 * sqrt(p[i] (1 - p[i]))) ∈ (1, ∞) here p is the minor allele frequency computed by maf() in SnpArrays.

source

Missing docstring.

Missing docstring for initialize_beta!. Check Documenter's build log for details.